requirements.txt
fastapi
uvicorn
openai
langchain
faiss-cpu
sentence-transformers
duckdb
pandas
numpy
scikit-learn
matplotlib
networkx
beautifulsoup4
lxml
pyyaml
aiofiles
python-docx
python-pptx
openpyxl
reportlab
pypandoc
tiktoken
requests
neo4j
selenium
python-dotenv